refactor: 타입 별 신규 파일 생성 및 APIProvider change 메서드 리네이밍
