+++ BoxOffice/Controller/ViewController.swift
@@ -8,6 +8,7 @@
 import UIKit
 
 class ViewController: UIViewController {
+    
     override func viewDidLoad() {
         super.viewDidLoad()
     }
