+++ BoxOffice/Resource/Assets.xcassets/Contents.json
